+++ .husky/pre-commit
@@ -1,4 +1,4 @@
 #!/bin/sh
 . "$(dirname "$0")/_/husky.sh"
 
-yarn pathpida && git add 'src/lib/$path.ts' && yarn lint-staged
+yarn pathpida && git add 'src/utils/$path.ts' && yarn lint-staged
